+++ Bounce.h
@@ -41,13 +41,13 @@ public:
 	// Updates the pin
 	// Returns 1 if the state changed
 	// Returns 0 if the state did not change
-  int update(); 
+  int16_t  update(); 
 	// Forces the pin to signal a change (through update()) in X milliseconds 
 	// even if the state does not actually change
 	// Example: press and hold a button and have it repeat every X milliseconds
   void rebounce(unsigned long interval); 
 	// Returns the updated pin state
-  int read();
+  int16_t read();
 	// Sets the stored pin state
   void write(int new_state);
     // Returns the number of milliseconds the pin has been in the current state
@@ -58,8 +58,8 @@ public:
 	bool fallingEdge();
   
 protected:
-  int debounce();
-  unsigned long  previous_millis, interval_millis, rebounce_millis;
+  uint16_t debounce();
+  uint16_t  previous_millis, interval_millis, rebounce_millis;
   uint8_t state;
   uint8_t pin;
   uint8_t stateChanged;
